arquiteture revised
