+++ src/services/front-faced/index.js
@@ -2,6 +2,7 @@
 import Koa from 'koa'
 import bodyParser from 'koa-bodyparser'
 import { Sockend } from 'cote'
+import { getAvailablePort } from 'utils'
 import { createServer } from 'http'
 import socket from 'socket.io'
 import router from './routes'
@@ -16,7 +17,7 @@ app.use(bodyParser())
 app.use(router.routes())
 app.use(router.allowedMethods())
 
-server.listen(5001);
+getAvailablePort(5001).then(port => server.listen(port))
 
 new Sockend(io, {
     name: 'front-faced sockend server'
