+++ yarn-error.log
@@ -1,5 +1,5 @@
 Arguments: 
-  /Users/badicomac/.nvm/versions/node/v8.10.0/bin/node /usr/local/bin/yarn test __tests__/services/front-faced/routes
+  /Users/badicomac/.nvm/versions/node/v8.10.0/bin/node /usr/local/bin/yarn add .
 
 PATH: 
   /Users/badicomac/.nvm/versions/node/v8.10.0/bin:/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin
@@ -25,7 +25,8 @@ npm manifest:
       "service-start": "BABEL_ENV=dev nodemon -e js,jsx,md,mustache --exec babel-node -- ",
       "migrate": "rethinkdb-migrate",
       "start": "node $OUTPUT_DIR/server/server",
-      "test": "jest --colors -b --verbose",
+      "test": "jest --colors -b --verbose --forceExit",
+      "aproval-test": "jest --colors -b --forceExit",
       "postinstall": "if [ \"$NODE_ENV\" = production ] ; then npm run prod-build ; fi",
       "prod-build-client": "webpack --config ./webpack/config.js",
       "prod-build-server": "babel ./src -d $OUTPUT_DIR/server --source-maps",
@@ -144,7 +145,7 @@ npm manifest:
     },
     "husky": {
       "hooks": {
-        "pre-push": "npm run test"
+        "pre-push": "npm run aproval-test"
       }
     }
   }
@@ -8864,17 +8865,14 @@ Lockfile:
     resolved "https://registry.yarnpkg.com/yeast/-/yeast-0.1.2.tgz#008e06d8094320c372dbc2f8ed76a0ca6c8ac419"
 
 Trace: 
-  Error: Command failed.
-  Exit code: 1
-  Command: sh
-  Arguments: -c jest --colors -b --verbose __tests__/services/front-faced/routes
-  Directory: /Users/badicomac/Repositorios/beeChallenge
-  Output:
-  
-      at ProcessTermError.MessageError (/usr/local/lib/node_modules/yarn/lib/cli.js:186:110)
-      at new ProcessTermError (/usr/local/lib/node_modules/yarn/lib/cli.js:226:113)
-      at ChildProcess.<anonymous> (/usr/local/lib/node_modules/yarn/lib/cli.js:30281:17)
-      at emitTwo (events.js:126:13)
-      at ChildProcess.emit (events.js:214:7)
-      at maybeClose (internal/child_process.js:925:16)
-      at Process.ChildProcess._handle.onexit (internal/child_process.js:209:5)
+  Error: Received malformed response from registry for undefined. The registry may be down.
+      at new MessageError (/usr/local/lib/node_modules/yarn/lib/cli.js:186:110)
+      at Function.<anonymous> (/usr/local/lib/node_modules/yarn/lib/cli.js:51419:15)
+      at Generator.next (<anonymous>)
+      at step (/usr/local/lib/node_modules/yarn/lib/cli.js:98:30)
+      at /usr/local/lib/node_modules/yarn/lib/cli.js:116:14
+      at new Promise (<anonymous>)
+      at new F (/usr/local/lib/node_modules/yarn/lib/cli.js:23451:28)
+      at Function.<anonymous> (/usr/local/lib/node_modules/yarn/lib/cli.js:95:12)
+      at Function.findVersionInRegistryResponse (/usr/local/lib/node_modules/yarn/lib/cli.js:51461:19)
+      at NpmResolver.<anonymous> (/usr/local/lib/node_modules/yarn/lib/cli.js:51479:28)
